+++ lua/plugins/config/nvim_cmp.lua
@@ -27,14 +27,17 @@ local function cmp_config()
 			['<C-p>'] = cmp.mapping.select_prev_item(),
 			['<C-n>'] = cmp.mapping.select_next_item(),
 			['<C-d>'] = cmp.mapping.scroll_docs(-4),
-			['<C-f>'] = cmp.mapping.scroll_docs(4),
+			['<C-u>'] = cmp.mapping.scroll_docs(4),
 			['<C-Space>'] = cmp.mapping.complete(),
-			['<C-e>'] = cmp.mapping.close(),
+			['<C-e>'] = cmp.mapping({
+				i = cmp.mapping.abort(),
+				c = cmp.mapping.close(),
+			}),
 			['<C-y>'] = cmp.mapping.confirm({
 				behavior = cmp.ConfirmBehavior.Insert,
 				select = true,
 			}),
-			['<CR>'] = cmp.mapping.confirm({ select = true }),
+			-- ['<CR>'] = cmp.mapping.confirm({ select = true }),
 		},
 		snippet = {
 			expand = function(args)
@@ -72,10 +75,11 @@ local function cmp_config()
 	npairs.setup({
 		disable_filetype = { 'TelescopePrompt' },
 		check_ts = true,
+		map_cr = true,
 	})
 
-	local cmp_autopairs = require('nvim-autopairs.completion.cmp')
-	cmp.event:on('confirm_done', cmp_autopairs.on_confirm_done({ map_char = { tex = '' } }))
+	-- local cmp_autopairs = require('nvim-autopairs.completion.cmp')
+	-- cmp.event:on('confirm_done', cmp_autopairs.on_confirm_done({ map_char = { tex = '' } }))
 
 	local wk = require('which-key')